Extract PM quality API module
